% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_confints.R
\name{confint_median_hs}
\alias{confint_median_hs}
\title{Confidence interval method for the median by the method of
Hettmansperger & #' Sheather (1991)}
\usage{
confint_median_hs(x, conf.level = 0.95, x_is_sorted = FALSE,
  interpolate = TRUE)
}
\arguments{
\item{x}{vector of observations}

\item{conf.level}{A conf.level * 100% confidence interval is
computed}

\item{x_is_sorted}{Boolean (Default: FALSE) to safe sorting x, if
it is already sorted. This is merely for speed reasons in
situations where it is more efficient to only sort x once.}

\item{interpolate}{Boolean (Default: TRUE) stating whether to
interpolate the order statistics. If no interpolation is selected
then this is just the standard exact procedure based on the order
statistics. Note: This procedure is conservative (i.e. coverage
is usualler larger than the nominal conf.level and hence the
interval is actually in general too large).}
}
\value{
A vector of length two containing the lower and upper limit
  of the confidence interval
}
\description{
Confidence interval method for the median by the method of
Hettmansperger & #' Sheather (1991)
}
\details{
The interpolation procedure suggested by Hettmansperger
  and Sheather (1986) for the median is applied to the order
  statistic.
}
\references{
Hettmansperger TP and Sheather SJ (1986), Confidence
  intervals based on interpolated order statistics, Statistics and
  Probability Letters, 4, p. 75-79.
}

