% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flint.R
\docType{data}
\name{flint}
\alias{flint}
\title{Water Monitoring Sample from Flint, Michigan}
\format{A data frame with 71 rows and 2 variables:
\describe{
  \item{lead}{The measured lead concentration in the tap water (in ppb = mg/L)}
  \item{exclude}{Logical indicating if the observations was removed by the authorities or not. In total two observations were removed as explained in the references.}
}}
\usage{
data(flint)
}
\description{
Water Monitoring Sample from Flint, Michigan
}
\references{
Langkjær-Bain, R. (2017), The murky tale of Flint's deceptive water data. Significance, 14: 16–21. \url{http://onlinelibrary.wiley.com/doi/10.1111/j.1740-9713.2017.01016.x/full}.

Youtube video by Michigan Radio explaining the computation of the quantile \url{https://www.youtube.com/watch?v=9pql00zr700}
}
\keyword{data}
