% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_confints.R
\name{quantile_confint_exact}
\alias{quantile_confint_exact}
\title{Standard exact two-sided quantile confidence interval based on the
binomial distribution}
\usage{
quantile_confint_exact(x, p, conf.level = 0.95, x_is_sorted = FALSE,
  fix_interval = TRUE)
}
\arguments{
\item{x}{vector of observations}

\item{p}{quantile of interest, \eqn{0 \leq p \leq 1}{0 <= p <= 1}}

\item{conf.level}{A \code{conf.level} * 100% confidence interval is
computed}

\item{x_is_sorted}{Boolean (Default: FALSE) to safe sorting x, if
it is already sorted. This is merely for speed reasons in
situations where it is more efficient to only sort x once.}

\item{fix_interval}{Boolean (Default: TRUE) For the case with no
interpolation, try to extend interval upwards if coverage is too
little.}
}
\value{
A vector of length two containing the lower and upper limit
  of the confidence interval
}
\description{
Standard exact two-sided quantile confidence interval based on the
binomial distribution
}
\details{
This function is a pure call-through to the Nyblom function
  with \code{interpolate=FALSE}.
}
\examples{
set.seed(123)
x <- rnorm(25)
quantile_confint_exact(x=x, p=0.8, conf.level=0.95)
}
\references{
Nyblom J, Note in interpolated order statistics,
  Statistics and Probability Letters 14, p. 129-131.
}
