% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_confints.R
\name{quantile_confint_nyblom}
\alias{quantile_confint_nyblom}
\title{Two-sided quantile confidence interval based on interpolating the order
statistic as suggested in Nyblom (1991)}
\usage{
quantile_confint_nyblom(
  x,
  p,
  conf.level = 0.95,
  x_is_sorted = FALSE,
  interpolate = TRUE,
  fix_interval = TRUE
)
}
\arguments{
\item{x}{vector of observations}

\item{p}{quantile of interest, \eqn{0 \leq p \leq 1}{0 <= p <= 1}}

\item{conf.level}{A conf.level * 100\% confidence interval is
computed}

\item{x_is_sorted}{Boolean (Default: FALSE) indicating if \code{x}
is already sorted and, hence, it is not necessary to sort it again.
This is merely for speed reasons in situations where it is more efficient
to only sort x once.}

\item{interpolate}{Boolean (Default: TRUE) stating whether to
interpolate the order statistics. If no interpolation is selected
then this is just the standard exact procedure based on the order
statistics. Note: This procedure is conservative (i.e. coverage
is usualler larger than the nominal conf.level and hence the
interval is actually in general too large).}

\item{fix_interval}{Boolean (Default: TRUE) For the case with no
interpolation, try to extend interval upwards if coverage is too
little.}
}
\value{
A vector of length two containing the lower and upper limit
  of the confidence interval
}
\description{
Two-sided quantile confidence interval based on interpolating the order
statistic as suggested in Nyblom (1991)
}
\details{
The interpolation procedure suggested by Nyblom (1992),
  which extends work by Hettmansperger and Sheather (1986) for the
  median is applied to the order statistic.
}
\examples{
set.seed(123)
x <- rnorm(25)
quantile_confint_nyblom(x=x, p=0.8, conf.level=0.95, interpolate=TRUE)
}
\references{
Nyblom J, Note in interpolated order statistics,
  Statistics and Probability Letters 14, p. 129-131.
}
