% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{qci_coverage_one_sim}
\alias{qci_coverage_one_sim}
\title{Computing the coverage of different confidence interval methods for
quantiles by Monte Carlo integration.}
\usage{
qci_coverage_one_sim(qci_fun, n, rfunc = rnorm, qfunc = qnorm,
  p = 0.5, conf.level = 0.95, ...)
}
\arguments{
\item{qci_fun}{Function which given n, p and conf.level
computed a set of different confidence intervals. Should return a
matrix of dimension 2 x (no. of methods) which contains the lower
and upper bound of each confidence interval method.}

\item{n}{Size of the sample to generate in the simulation}

\item{rfunc}{Function for generating the samples}

\item{qfunc}{Quantile function for computing the true quantile}

\item{p}{The quantile of interest 0 <= p <= 1}

\item{conf.level}{conf.level * 100% two-sided confidence intervals
are computed}

\item{\dots}{Additional arguments passed to \code{rfunc} and \code{qfunc}}
}
\value{
A vector of Booleans of length (no. methods) stating if
  each method contains the true value or not
}
\description{
Computing the coverage of different confidence interval methods for
quantiles by Monte Carlo integration.
}
